Improve kernel robustness and fault/OOM flow
Large refactor and hardening across kernel and userspace to improve memory safety, fault handling, OOM behavior, capability revocation, and cleanup logic. Key changes:

- Expanded and rewrote the kernel robustness design doc (added migration plan, architectural invariants, fault/OOM/process state machine, lock hierarchy, callback isolation, and Portuguese translations/notes).
- Introduced typed ABI address/ range validation (UserVAddr/UserRange) and moved user addr validation into shared/abi.
- Split fault path into three layers: classify → admit (quota) → materialize; replaced boolean results with typed FaultResult/FaultError.
- Added per-process OOM model: snapshot-based victim selection, process-level kill (Dying/Dead states), deterministic teardown and OOM result types (OomResult).
- Implemented two-phase capability revocation with explicit RevokeReport and error reporting; forbid silent ignores of revoke results.
- Isolated callback invocation (clone-and-call outside locks) to prevent callback-in-lock issues and documented panic handling.
- Formalized lock acquisition hierarchy and removed panic/asserts from normal syscall/operational paths in favor of structured errors.
- Multiple kernel modules updated to follow the new contracts (mm/*, addrspace/vm/vma/policy, process/thread, interrupts, init_process, syscall surface) and shared ABI changes.
- Added userspace syscall VM shim (userspace/libs/syscall/src/vm.rs) and updated libc/graphics/syscall glue and apps accordingly.
- Minor build script and ancillary changes to reflect API/ABI modifications.

This commit is focused on defense-in-depth: explicit typing of failure modes, single source of truth for user addresses, process-scoped OOM handling, robust revocation semantics, and safer callback/lock semantics to prevent panics, leaks, and deadlocks.

+++ userspace/libs/libgui/src/surface.rs
@@ -83,7 +83,7 @@ impl Surface {
 
         if let Some(addr) = inner.shared.address() {
             let offset = (y * inner.shared.stride() + x) as usize * inner.shared.bytes_per_pixel();
-            let ptr = (addr + offset) as *const u32;
+            let ptr = ((addr as usize) + offset) as *const u32;
             let value = unsafe { ptr.read_volatile() };
             return Some(Color::rgb(
                 (value & 0xFF) as u8,
